+++ test/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TestQuaternion.java
@@ -130,7 +130,7 @@ public class TestQuaternion {
 	}
 
 	@Test
-	public void testSetType() {
+	public void testSetMessageType() {
 		empty.setMessageType("test");
 		assertEquals("test", empty.getMessageType());
 	}